removed sign out button from ProfileViewController and added signout functionality to SettingsViewController.
