+++ Nutriq/ViewControllers/HOME-ViewControllers/ProfileViewController.swift
@@ -8,7 +8,6 @@
 
 import UIKit
 import Firebase
-import GoogleSignIn
 
 class ProfileViewController: UIViewController {
     
@@ -22,18 +21,6 @@ class ProfileViewController: UIViewController {
         // Make the label transparent for a nice transition
         userEmailLabel.alpha = 0
         loadUserData()
-    }
-    
-
-    // MARK: - Sign Out Handling
-    
-    @objc func handleSignOut() {
-        let signOutAlertController = UIAlertController(title: nil, message: "Are you sure you want to sign out?", preferredStyle: .actionSheet)
-        signOutAlertController.addAction(UIAlertAction(title: "Sign Out", style: .destructive, handler: { (_) in
-            self.signOut()
-        }))
-        signOutAlertController.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: nil))
-        present(signOutAlertController, animated: true, completion: nil)
     }
     
     
@@ -52,44 +39,8 @@ class ProfileViewController: UIViewController {
         }
     }
     
-    // Sign user out - works for BOTH Google and email sign out
-    func signOut() {
-        GIDSignIn.sharedInstance().signOut() // Google sign out
-        do {
-            // TODO: - Before release, check to make sure that not checking which sign out is occuring will not impact the app in any way
-            try Auth.auth().signOut() // Email sign out
-            self.takeUserToUsernameScreen()
-        } catch let error {
-            print("Failed to sign out with error: ", error.localizedDescription)
-        }
-    }
-
-    @IBAction func signoutButtonPressed(_ sender: Any) {
-        handleSignOut()
-    }
-    
-    
-    func takeUserToUsernameScreen() {
-        let loginVC = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "LoginViewController")
-        self.segueToBottom()
-        UIApplication.topViewController()?.present(loginVC, animated: false, completion: nil)
-    }
-    
-    // Right-to-left segue animation
-    func segueToBottom() {
-        let transition = CATransition()
-        transition.duration = 0.5
-        transition.type = CATransitionType.push
-        transition.subtype = CATransitionSubtype.fromBottom
-        transition.timingFunction = CAMediaTimingFunction(name:CAMediaTimingFunctionName.easeInEaseOut)
-        self.view.window!.layer.add(transition, forKey: kCATransition)
-    }
-
-    
     /*
     // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         // Get the new view controller using segue.destination.
         // Pass the selected object to the new view controller.
